Extract build_context to Pipelines
